Refactor project structure, update imports, and improve Docker/local environment setup
