+++ alembic/env.py
@@ -5,10 +5,10 @@ from sqlalchemy import pool
 
 from alembic import context
 
-from models.article import Article
-from models.source import Source
-from database import Base
-from config import settings
+from app.models.article import Article
+from app.models.source import Source
+from app.database import Base
+from app.config import settings
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
