+++ app/database.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from config import settings
+from app.config import settings
 
 # pick the URL based on ENV
 engine = create_engine(settings.SQLALCHEMY_DATABASE_URL, echo=True)
