+++ app/models/__init__.py
@@ -1,0 +1,4 @@
+from app.models.article import Article
+from app.models.user import User
+from app.models.bookmark import Bookmark
+from app.models.source import Source
