+++ app/models/bookmark.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, ForeignKey
 from sqlalchemy.orm import relationship
-from database import Base
+from app.database import Base
 
 class Bookmark(Base):
     __tablename__ = "bookmarks"
