+++ app/models/source.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
-from database import Base
+from app.database import Base
 
 class Source(Base):
     __tablename__ = "sources"
