+++ app/routers/articles.py
@@ -2,11 +2,11 @@ from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from typing import List
 
-from database import get_db
-from models.article import Article as ArticleModel
-from schemas.article import ArticleRead
+from app.database import get_db
+from app.models.article import Article as ArticleModel
+from app.schemas.article import ArticleRead
 
-router = APIRouter(prefix="/articles", tags=["articles"])
+router = APIRouter(tags=["Articles"])
 
 @router.get("/", response_model=List[ArticleRead])
 def get_articles(db: Session = Depends(get_db), limit: int = 20):
