--- available_sources.csv
@@ -1,54 +1,0 @@
-Code,Name,Category,Country,Language,URL
-skynews,Sky News - Politics,politics,us,en,https://news.sky.com/
-bbc,BBC News - UK Politics,politics,gb,en,http://bbc.com/news/politics/
-foxnews,FOX News - Politics,politics,us,en,https://foxnews.com
-foreignpolicy,Foreign Policy,general,us,en,blog.foreignpolicy.com
-businesstoday,Business Today | Latest Stock Market And Economy News India,business,in,en,http://www.businesstoday.in/
-bloomberg,Bloomberg | Latest And Live Business,business,us,en,https://bloomberg.com/
-financialpost,Financial Post | Canada Business News,business,us,en,https://business.financialpost.com/
-bworldonline,Business World,business,us,en,https://www.bworldonline.com/
-businesstravelnews,Business Travel News,business,us,en,https://www.businesstravelnews.com/
-libn,Long Island Business News,business,us,en,https://libn.com/
-iotbusinessnews,Iot Business News,business,us,en,https://iotbusinessnews.com/
-business,business,general,us,en,business.newsfactor.com
-yahoo,Yahoo | Business Finance,business,us,en,https://finance.yahoo.com/news/
-mjbizdaily,Marijuana Business Daily,business,us,en,https://mjbizdaily.com/
-businessnews,Business News.com.au,business,au,en,https://www.businessnews.com.au/
-thehindubusinessline,Business Line,business,us,en,https://www.thehindubusinessline.com/
-zeebiz,Zee Business,business,us,en,https://www.zeebiz.com/
-canadianbusiness,Canadian Business | Business News,business,us,en,https://www.canadianbusiness.com/business-news/
-businessnews-ng,Nigeria Business News,business,us,en,http://businessnews.com.ng/
-cnbc,Cnbc | World Business News Leader,business,us,en,http://www.cnbc.com/
-google-news,Google News Business US,business,us,en,https://news.google.com
-arabian-business,Arabian Business,general,ae,en,http://arabianbusiness.com/
-markets,markets,general,us,en,www.markets.com
-finance-commerce,finance-commerce,general,us,en,finance-commerce.com
-scidev,Science &amp; Technology For Global Development News &amp; Analysis - Scidev.net,science,us,en,https://www.scidev.net/
-scitechdaily,Scitechdaily | Science Space And Technology News 2017,science,us,en,http://scitechdaily.com/
-phys,Phys.org - News And Articles On Science And Technology,science,us,en,https://phys.org/
-scienceandtechnologyresearchnews,Science And Technology Research News,science,us,en,http://www.scienceandtechnologyresearchnews.com/
-popsci,Popular Science Magazine | Science Space And New Technology,science,us,en,https://www.popsci.com/
-cnn,CNN Technology,technology,us,en,http://cnn.com/
-tech,tech,general,us,en,tech.blorge.com
-wwos,Wide World Of Sports,sports,au,en,https://wwos.nine.com.au/
-worldinsport,World In Sport,sports,us,en,https://worldinsport.com/
-haaretz,Haaretz - Jewish World,general,il,en,http://www.haaretz.com
-sportsration,Sportsration: Nigeria Soccer | World Sports,sports,us,en,https://sportsration.com/
-sportskanazee,Sportskanazee | Stories From The Sports World,sports,us,en,https://www.sportskanazee.com/
-ibtimes,International Business Times Au,business,au,en,https://www.ibtimes.com.au/
-global-news,Global News,general,ca,en,http://www.globalnews.ca/
-global-times,Global Times,general,sl,en,http://www.globaltimes-sl.org/
-net-security,net-security,general,us,en,www.net-security.org
-nature,Nature,science,us,en,blogs.nature.com
-livescience,Live Science: The Most Interesting Articles,science,us,en,https://www.livescience.com/
-xplorelifescience,Xplore Life Science,science,us,en,https://www.xplorelifescience.com/
-independentsciencenews,Independent Science News | Food Health And Agriculture Bioscience News,science,us,en,https://www.independentsciencenews.org/
-sciworthy,Science For Everyone: The Latest Research Papers Explained,science,us,en,https://sciworthy.com/
-advancedsciencenews,Advanced Science News,science,us,en,https://www.advancedsciencenews.com/
-eurekalert,Eurekalert! Science News,science,us,en,https://www.eurekalert.org/
-sciencenewsforstudents,Science News For Students,science,us,en,https://www.sciencenewsforstudents.org/
-sciencemag,Science | Aaas,science,us,en,http://www.sciencemag.org/
-undark,Truth Beauty Science. - Undark,science,us,en,https://undark.org/
-sciencenewsjournal,Science News Journal,science,us,en,http://sciencenewsjournal.com/
-eos,Eos: Earth And Space Science News,science,us,en,https://eos.org/
-crime-research,crime-research,general,us,en,www.crime-research.org
