+++ docker-compose.yml
@@ -2,7 +2,7 @@ services:
   web:
     build: .
     container_name: aifeelnews-api
-    command: uvicorn main:app --host 0.0.0.0 --port 8000 --reload
+    command: uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload
     volumes:
       - .:/app
     ports:
