--- models/__init__.py
@@ -1,4 +1,0 @@
-from .article import Article
-from .user import User
-from .bookmark import Bookmark
-from .source import Source
